requirements.txt
argparse-formatter>=1.4
colorama>=0.4.4
construct~=2.10.67
fusion-engine-client==1.17.0
gpstime>=0.6.2
pynmea2~=1.18.0
pyserial~=3.5
urllib3>=1.21.1
websockets>=10.1

# Required for SEGGER RTT client use only.
psutil>=5.9.4

# Required to avoid errors when downloading packages from github below.
wheel>=0.37.1

# Note: Using the P1 fork of ntripstreams until fixes are mainlined.
https://github.com/PointOneNav/ntripstreams/archive/c26605710a53a1ebe1a16310565b5605e77228c1.zip#egg=ntripstreams
#ntripstreams~=0.1.12
